chore: introduce service_loader module
